backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymysql==1.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
httpx==0.25.1